+++ src/components/CheckboxGroup/CheckboxGroup.stories.tsx
@@ -3,7 +3,7 @@ import { Button, TextArea } from "components";
 import { Checkbox, CheckboxProps } from "components/Checkbox/Checkbox";
 import { FormEvent, useCallback, useState } from "react";
 import { CheckboxGroup, CheckboxGroupProps } from "./CheckboxGroup";
-import { checkboxes, disabledCheckboxes, readonlyCheckboxes } from "./helpers";
+import { checkboxes, disabledCheckboxes, readonlyCheckboxes } from "./testHelpers";
 
 export default {
   title: "Components/Checkbox Group",
