+++ src/components/CheckboxGroup/testHelpers.tsx
